#5 Création de compte : Best case

+++ src/main/java/org/sqli/authentification/entitie/GroupRepository.java
@@ -3,7 +3,7 @@ package org.sqli.authentification.entitie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.sqli.authentification.dao.Group;
 
-public interface GroupRepository extends JpaRepository<Group, Integer> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
    // Long getGroupById(int toIntExact);
-    //int getGroupById(int id);
+    //Group getGroupById(Long id);
 }
